added PDF previews
